Backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
pytest
pytest-django
python-dotenv
requests
openpyxl
psycopg2-binary
gunicorn
